Migrate wasm-bindgen `char` test to `wasm`
