+++ tests/all/main.rs
@@ -4,7 +4,6 @@ extern crate wasm_bindgen_test_project_builder as project_builder;
 
 use project_builder::{project, run};
 
-mod char;
 mod classes;
 mod closures;
 mod comments;
