+++ tests/wasm/main.rs
@@ -5,5 +5,6 @@ extern crate wasm_bindgen_test;
 extern crate wasm_bindgen;
 
 pub mod api;
+pub mod char;
 pub mod option;
 pub mod optional_primitives;
